icon

+++ code/readingfromfilelnx_v3.py
@@ -3,11 +3,13 @@ from datetime import date
 from time import sleep
 
 #ver14 new
-filelist = [ f for f in os.listdir('/home/pi/icons') if (f.endswith('ver.py'))]
-for f in filelist:
-	os.remove('/home/pi/icons/'+f) 
-if filelist:
-	os.system('sudo mv ' +'/home/pi/icons/ver2.py '+ '/home/pi/icons/ver.py')
+filelist1 = [ f for f in os.listdir('/home/pi/icons') if (f.endswith('ver2.py'))]
+if filelist1:
+	filelist2 = [ f for f in os.listdir('/home/pi/icons') if (f.endswith('ver.py'))]
+	for f in filelist2:
+		os.remove('/home/pi/icons/'+f) 
+	if filelist2:
+		os.system('sudo mv ' +'/home/pi/icons/ver2.py '+ '/home/pi/icons/ver.py')
 
 today = date.today()
 day = today.day			#testing
